Implemented user store and all user related actions

+++ app/src/main/java/me/stammberger/starcitizencompact/models/user/Performance.java
@@ -10,7 +10,7 @@ public class Performance {
 
     @SerializedName("network_io_time")
     @Expose
-    public Integer networkIoTime;
+    public Double networkIoTime;
     @SerializedName("processing_time")
     @Expose
     public Double processingTime;
